requirements.txt
boto3
pyyaml
os
json
gzip
datetime